Re-instate simple search api that takes in pipeline. Use schema via app context.
